+++ model/src/main/scala/org/broadinstitute/dsde/workbench/model/UserInfo.scala
@@ -2,4 +2,5 @@ package org.broadinstitute.dsde.workbench.model
 
 import akka.http.scaladsl.model.headers.OAuth2BearerToken
 
-case class UserInfo(accessToken: OAuth2BearerToken, user: WorkbenchUser, tokenExpiresIn: Long)
+case class UserInfo(accessToken: OAuth2BearerToken, userId: WorkbenchUserId, userEmail: WorkbenchEmail, tokenExpiresIn: Long)
+final case class CreateWorkbenchUserAPI(id: WorkbenchUserId, googleSubjectId: GoogleSubjectId, email: WorkbenchEmail)