+++ model/CHANGELOG.md
@@ -3,7 +3,7 @@
 This file documents changes to the `workbench-model` library, including notes on how to upgrade to new versions.
 
 ## 0.12
-- Add `LdapUser`
+- Add `googleSubjectId` as a separate field to `WorkbenchUser`
 
 ## 0.11